all search integration and unit test finished

+++ server/src/main/java/com/culturaloffers/maps/repositories/CulturalOfferRepository.java
@@ -22,7 +22,7 @@ public interface CulturalOfferRepository extends JpaRepository<CulturalOffer, In
     List<CulturalOffer> findBySubtypeContaining(String name);
 
     @Query("Select distinct(offer) from CulturalOffer offer join offer.subscribers subscribers where size(subscribers) >= :amount")
-    List<CulturalOffer> findBySubscriberAmountGreaterEq(Long amount);
+    List<CulturalOffer> findBySubscriberAmountGreaterEq(Integer amount);
 
     @Query("Select distinct(offer) from CulturalOffer offer join offer.userGrades grades where size(grades) > 0")
     List<CulturalOffer> findGraded();
